feat: 点餐助手可绑定的收银机列表
时间戳使用int64类型，可避免转换

+++ main/app/dto/req/bind_record.go
@@ -5,7 +5,7 @@ type AddBindRecordReq struct {
 	Brand            string `json:"brand"`              // 品牌
 	Source           string `json:"source"`             // 来源：cashier、tablet、kitchen、assistant
 	FinallyLoginUuid uint64 `json:"finally_login_uuid"` // 最后登录的员工ID
-	FinallyLoginTime int    `json:"finally_login_time"` // 最后登录的时间
+	FinallyLoginTime int64  `json:"finally_login_time"` // 最后登录的时间
 	CompanyUuid      uint64 `json:"company_uuid"`       // 公司ID
 	PrintPortUuid    uint64 `json:"print_port_uuid"`    // 打印端口ID
 	UserAgent        string `json:"user_agent"`         // 用户代理
